requirements.txt
numpy
pandas
matplotlib
seaborn
nltk
scikit-learn
pypdf
streamlit
python-docx
regex 
